+++ trixnity-client/src/commonMain/kotlin/de/connect2x/trixnity/client/room/RoomUpgradeHandler.kt
@@ -76,7 +76,6 @@ class RoomUpgradeHandler(
                 }.onJoin { }
                 launch {
                     delay(initialDelay)
-
                 }.onJoin { }
             }.also {
                 currentCoroutineContext().cancelChildren()